+++ lib/jellyfish/normalized_path.rb
@@ -4,7 +4,7 @@ require 'uri'
 
 
 module Jellyfish
-  module UnescapePath
+  module NormalizedPath
     def path_info
       URI.decode_www_form_component(super, Encoding.default_external)
     end